+++ src/gui/graphics/item_data.hpp
@@ -6,7 +6,8 @@ enum ItemData
 {
     NodePointer,
     HandleRole,
-    GradientStopIndex
+    GradientStopIndex,
+    AssociatedProperty
 };
 
 } // namespace graphics