+++ src/gui/widgets/shape_style/fill_style_widget.hpp
@@ -1,15 +1,17 @@
 #pragma once
 #include "color_selector.hpp"
 #include "model/shapes/fill.hpp"
+#include "utils/pseudo_mutex.hpp"
 
 class FillStyleWidget : public ColorSelector
 {
 public:
     FillStyleWidget(QWidget* parent = nullptr);
 
-    void set_shape(model::Fill* target);
-
+    void set_shape(model::Fill* target, int gradient_stop = 0);
     model::Fill* shape() const;
+
+    void set_gradient_stop(model::Styler* styler, int index);
 
 private:
     void update_from_target();
@@ -25,6 +27,7 @@ private slots:
 
 private:
     model::Fill* target = nullptr;
-    bool updating = false;
+    utils::PseudoMutex updating;
+    int stop = 0;
 };