+++ CHANGELOG.md
@@ -6,6 +6,10 @@
     * Added gradient support
     * Option to exclude some layers from being rendered / exported
     * Freehand drawing tool
+    * Interactive "add node" with the edit tool
+    * Bezier molding, click and drag on a curve to edit it, without using handles
+    * "Dissolve nodes", remove bezier nodes while trying to keep the old shape
+    * Right-clicking on a handle with the select or edit tools will display a context menu for it
 * Bug Fixes:
     * Exporting the document no longer alters the document save state
     * Fixed crash on Lottie export